requirements.txt
Flask==3.0.2
gunicorn==21.2.0
pandas==2.2.2
numpy==1.26.4
yfinance==0.2.40
APScheduler==3.10.4
python-dotenv==1.0.1
twilio==9.2.3
requests==2.32.3
